authenticate paint by sending the userName and token in request
the token is sent to the user in email, it can be input in the main UI
(which is auto-populated when clicking in the email), and is sent (for
now as RequestParam) to the Controller, which validates it.

token expires after 30min, and is not removed from TokenService when
validated.